update 1
added flies to boss heart, also changed everythings x position to be
represented by a float instead of an int because that makes low
velocities be less clunky. flies still not completely coded properly

+++ core/src/com/mygdx/game2/HealthCrystal.java
@@ -34,7 +34,7 @@ public class HealthCrystal extends PlayerItem {
     private int frameCount = 0;
     private int framesSwitch = 6; // TODO: idea, could possibly have framesswitch go up as time goes on to seem like spinning is slowing
 
-    public HealthCrystal(int x, int y, float velX, float velY, int width, int height) {
+    public HealthCrystal(float x, float y, float velX, float velY, int width, int height) {
         super(x,y,velX,velY,width,height);
         destroyed = false;
         type = "Health";
